MVC update (read description)
- The API is not functioning yet to retrieve book and author data.
- The application is consuming excessive device resources, and the cause is not yet identified.
