+++ lib/src/constants/text_strings.dart
@@ -40,3 +40,10 @@ const String tPenulisDicariTitle = 'Penulis Paling Dicari';
 const String tBukuAjarTitle = 'Buku Ajar';
 const String tJournalTitle = 'Journal';
 const String tProceedingTitle = 'Proceeding';
+
+// Favorite
+const String tFavoriteTitle = 'Favorit';
+const String tFavoriteBody1 = 'Simpan daftar buku \nuntuk kamu baca';
+const String tFavoriteBody2 =
+    'Untuk menambahkan buku, klik ikon\n wishlist yang ada pada detail buku';
+const String tFavoriteButtonText = 'Temukan Buku';
