+++ lib/src/features/home/model/beranda_model.dart
@@ -1,0 +1,7 @@
+class BerandaModel {
+  List<Map<String, dynamic>> userDataList = [];
+  List<Map<String, dynamic>> latestBooks = [];
+  List<Map<String, dynamic>> generalBooks = [];
+  List<Map<String, dynamic>> journals = [];
+  List<Map<String, dynamic>> proceedings = [];
+}
