+++ lib/src/features/home/view/books/allbook_widget.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+import '../../../../constants/text_strings.dart';
+
+Widget dashboardLihatSemua(BuildContext context) {
+  return Padding(
+    padding: const EdgeInsets.only(right: 24),
+    child: Text(
+      tLihatSemuaTitle,
+      style: Theme.of(context).textTheme.labelSmall,
+    ),
+  );
+}
