+++ lib/src/features/home/view/books/buku_terbaru_widget.dart
@@ -1,0 +1,31 @@
+import 'package:bookstore/src/features/home/controller/beranda_controller.dart';
+import 'package:flutter/material.dart';
+
+import 'allbook_widget.dart';
+import 'books_scroll.dart';
+
+Widget dashboardTerbaru(BuildContext context, BerandaController controller) {
+  var latest = controller.model.latestBooks;
+  return Column(
+    children: [
+      Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          const Text(
+            "Buku Terbaru",
+            style: TextStyle(
+              fontSize: 20,
+              fontWeight: FontWeight.w500,
+            ),
+          ),
+          dashboardLihatSemua(context),
+        ],
+      ),
+      const SizedBox(
+        height: 12,
+      ),
+      // buku terbaru
+      booksHorizontalScroll(context, latest, controller),
+    ],
+  );
+}
