+++ lib/src/utils/theme/widget_theme/text_theme.dart
@@ -5,6 +5,18 @@ class TTextTheme {
   TTextTheme._(); // menghindari penggunaan instance
 
   static TextTheme lightTextTheme = TextTheme(
+      headlineSmall: GoogleFonts.roboto(
+        color: Colors.black,
+        fontSize: 20,
+        fontWeight: FontWeight.w500,
+      ),
+
+      // Digunakan pada tombol Lihat semua
+      /*labelSmall: GoogleFonts.inter(
+        // color: Color(0xFF7A7A7A),
+        color: Colors.red,
+        fontSize: 14,
+      ),*/
       headline2: GoogleFonts.montserrat(
         color: Colors.black87,
         fontSize: 16,
